renames the ducks and fixes test
