+++ frontend/amundsen_application/static/js/ducks/popularResources/reducer.ts
