+++ frontend/amundsen_application/static/js/ducks/popularResources/types.ts
